feat: move process functions to be used outside of this repository + ProcessSource
