+++ cmd/tlin/main.go
@@ -9,7 +9,6 @@ import (
 	"go/parser"
 	"go/token"
 	"os"
-	"path/filepath"
 	"sort"
 	"strings"
 	"time"
@@ -18,8 +17,8 @@ import (
 	"github.com/gnoswap-labs/tlin/internal"
 	"github.com/gnoswap-labs/tlin/internal/analysis/cfg"
 	"github.com/gnoswap-labs/tlin/internal/fixer"
-	"github.com/gnoswap-labs/tlin/internal/lints"
 	tt "github.com/gnoswap-labs/tlin/internal/types"
+	"github.com/gnoswap-labs/tlin/lint"
 	"go.uber.org/zap"
 )
 
@@ -42,11 +41,6 @@ type Config struct {
 	ConfidenceThreshold  float64
 }
 
-type LintEngine interface {
-	Run(filePath string) ([]tt.Issue, error)
-	IgnoreRule(rule string)
-}
-
 func main() {
 	logger, _ := zap.NewProduction()
 	defer logger.Sync()
@@ -56,7 +50,7 @@ func main() {
 	ctx, cancel := context.WithTimeout(context.Background(), config.Timeout)
 	defer cancel()
 
-	engine, err := internal.NewEngine(".")
+	engine, err := internal.NewEngine(".", nil)
 	if err != nil {
 		logger.Fatal("Failed to initialize lint engine", zap.Error(err))
 	}
@@ -133,8 +127,8 @@ func runWithTimeout(ctx context.Context, f func()) {
 	}
 }
 
-func runNormalLintProcess(ctx context.Context, logger *zap.Logger, engine LintEngine, paths []string, jsonOutput string) {
-	issues, err := processFiles(ctx, logger, engine, paths, processFile)
+func runNormalLintProcess(ctx context.Context, logger *zap.Logger, engine lint.LintEngine, paths []string, jsonOutput string) {
+	issues, err := lint.ProcessFiles(ctx, logger, engine, paths, lint.ProcessFile)
 	if err != nil {
 		logger.Error("Error processing files", zap.Error(err))
 		os.Exit(1)
@@ -148,8 +142,8 @@ func runNormalLintProcess(ctx context.Context, logger *zap.Logger, engine LintEn
 }
 
 func runCyclomaticComplexityAnalysis(ctx context.Context, logger *zap.Logger, paths []string, threshold int, jsonOutput string) {
-	issues, err := processFiles(ctx, logger, nil, paths, func(_ LintEngine, path string) ([]tt.Issue, error) {
-		return processCyclomaticComplexity(path, threshold)
+	issues, err := lint.ProcessFiles(ctx, logger, nil, paths, func(_ lint.LintEngine, path string) ([]tt.Issue, error) {
+		return lint.ProcessCyclomaticComplexity(path, threshold)
 	})
 	if err != nil {
 		logger.Error("Error processing files for cyclomatic complexity", zap.Error(err))
@@ -192,11 +186,11 @@ func runCFGAnalysis(_ context.Context, logger *zap.Logger, paths []string, funcN
 	}
 }
 
-func runAutoFix(ctx context.Context, logger *zap.Logger, engine LintEngine, paths []string, dryRun bool, confidenceThreshold float64) {
+func runAutoFix(ctx context.Context, logger *zap.Logger, engine lint.LintEngine, paths []string, dryRun bool, confidenceThreshold float64) {
 	fix := fixer.New(dryRun, confidenceThreshold)
 
 	for _, path := range paths {
-		issues, err := processPath(ctx, logger, engine, path, processFile)
+		issues, err := lint.ProcessPath(ctx, logger, engine, path, lint.ProcessFile)
 		if err != nil {
 			logger.Error("error processing path", zap.String("path", path), zap.Error(err))
 			continue
@@ -207,64 +201,6 @@ func runAutoFix(ctx context.Context, logger *zap.Logger, engine LintEngine, path
 			logger.Error("error fixing issues", zap.String("path", path), zap.Error(err))
 		}
 	}
-}
-
-func processFiles(ctx context.Context, logger *zap.Logger, engine LintEngine, paths []string, processor func(LintEngine, string) ([]tt.Issue, error)) ([]tt.Issue, error) {
-	var allIssues []tt.Issue
-	for _, path := range paths {
-		issues, err := processPath(ctx, logger, engine, path, processor)
-		if err != nil {
-			logger.Error("Error processing path", zap.String("path", path), zap.Error(err))
-			return nil, err
-		}
-		allIssues = append(allIssues, issues...)
-	}
-
-	return allIssues, nil
-}
-
-func processPath(_ context.Context, logger *zap.Logger, engine LintEngine, path string, processor func(LintEngine, string) ([]tt.Issue, error)) ([]tt.Issue, error) {
-	info, err := os.Stat(path)
-	if err != nil {
-		return nil, fmt.Errorf("error accessing %s: %w", path, err)
-	}
-
-	var issues []tt.Issue
-	if info.IsDir() {
-		err = filepath.Walk(path, func(filePath string, fileInfo os.FileInfo, err error) error {
-			if err != nil {
-				return err
-			}
-			if !fileInfo.IsDir() && hasDesiredExtension(filePath) {
-				fileIssues, err := processor(engine, filePath)
-				if err != nil {
-					logger.Error("Error processing file", zap.String("file", filePath), zap.Error(err))
-				} else {
-					issues = append(issues, fileIssues...)
-				}
-			}
-			return nil
-		})
-		if err != nil {
-			return nil, fmt.Errorf("error walking directory %s: %w", path, err)
-		}
-	} else if hasDesiredExtension(path) {
-		fileIssues, err := processor(engine, path)
-		if err != nil {
-			return nil, err
-		}
-		issues = append(issues, fileIssues...)
-	}
-
-	return issues, nil
-}
-
-func processFile(engine LintEngine, filePath string) ([]tt.Issue, error) {
-	return engine.Run(filePath)
-}
-
-func processCyclomaticComplexity(path string, threshold int) ([]tt.Issue, error) {
-	return lints.DetectHighCyclomaticComplexity(path, threshold)
 }
 
 func printIssues(logger *zap.Logger, issues []tt.Issue, jsonOutput string) {
@@ -287,7 +223,7 @@ func printIssues(logger *zap.Logger, issues []tt.Issue, jsonOutput string) {
 				logger.Error("Error reading source file", zap.String("file", filename), zap.Error(err))
 				continue
 			}
-			output := formatter.GenetateFormattedIssue(fileIssues, sourceCode)
+			output := formatter.GenerateFormattedIssue(fileIssues, sourceCode)
 			fmt.Println(output)
 		}
 	} else {
@@ -308,8 +244,4 @@ func printIssues(logger *zap.Logger, issues []tt.Issue, jsonOutput string) {
 			return
 		}
 	}
-}
-
-func hasDesiredExtension(path string) bool {
-	return filepath.Ext(path) == ".go" || filepath.Ext(path) == ".gno"
 }
