+++ cmd/tlin/main_test.go
@@ -29,6 +29,11 @@ func (m *mockLintEngine) Run(filePath string) ([]types.Issue, error) {
 	return args.Get(0).([]types.Issue), args.Error(1)
 }
 
+func (m *mockLintEngine) RunSource(source []byte) ([]types.Issue, error) {
+	args := m.Called(source)
+	return args.Get(0).([]types.Issue), args.Error(1)
+}
+
 func (m *mockLintEngine) IgnoreRule(rule string) {
 	m.Called(rule)
 }
@@ -98,116 +103,6 @@ func TestParseFlags(t *testing.T) {
 			assert.Equal(t, tt.expected.JsonOutput, config.JsonOutput)
 		})
 	}
-}
-
-func TestProcessFile(t *testing.T) {
-	t.Parallel()
-	expectedIssues := []types.Issue{
-		{
-			Rule:     "test-rule",
-			Filename: "test.go",
-			Start:    token.Position{Filename: "test.go", Offset: 0, Line: 1, Column: 1},
-			End:      token.Position{Filename: "test.go", Offset: 10, Line: 1, Column: 11},
-			Message:  "Test issue",
-		},
-	}
-	mockEngine := setupMockEngine(expectedIssues, "test.go")
-
-	issues, err := processFile(mockEngine, "test.go")
-
-	assert.NoError(t, err)
-	assert.Equal(t, expectedIssues, issues)
-	mockEngine.AssertExpectations(t)
-}
-
-func TestProcessPath(t *testing.T) {
-	t.Parallel()
-	logger, _ := zap.NewProduction()
-	ctx := context.Background()
-
-	tempDir, err := os.MkdirTemp("", "test")
-	assert.NoError(t, err)
-	defer os.RemoveAll(tempDir)
-
-	paths := createTempFiles(t, tempDir, "test1.go", "test2.go")
-
-	expectedIssues := []types.Issue{
-		{
-			Rule:     "rule1",
-			Filename: paths[0],
-			Start:    token.Position{Filename: paths[0], Offset: 0, Line: 1, Column: 1},
-			End:      token.Position{Filename: paths[0], Offset: 10, Line: 1, Column: 11},
-			Message:  "Test issue 1",
-		},
-		{
-			Rule:     "rule2",
-			Filename: paths[1],
-			Start:    token.Position{Filename: paths[1], Offset: 0, Line: 1, Column: 1},
-			End:      token.Position{Filename: paths[1], Offset: 10, Line: 1, Column: 11},
-			Message:  "Test issue 2",
-		},
-	}
-
-	mockEngine := new(mockLintEngine)
-	mockEngine.On("Run", paths[0]).Return([]types.Issue{expectedIssues[0]}, nil)
-	mockEngine.On("Run", paths[1]).Return([]types.Issue{expectedIssues[1]}, nil)
-
-	issues, err := processPath(ctx, logger, mockEngine, tempDir, processFile)
-
-	assert.NoError(t, err)
-	assert.Len(t, issues, 2)
-	assert.Contains(t, issues, expectedIssues[0])
-	assert.Contains(t, issues, expectedIssues[1])
-	mockEngine.AssertExpectations(t)
-}
-
-func TestProcessFiles(t *testing.T) {
-	t.Parallel()
-	logger, _ := zap.NewProduction()
-	ctx := context.Background()
-
-	tempDir, err := os.MkdirTemp("", "test")
-	assert.NoError(t, err)
-	defer os.RemoveAll(tempDir)
-
-	paths := createTempFiles(t, tempDir, "test1.go", "test2.go")
-
-	expectedIssues := []types.Issue{
-		{
-			Rule:     "rule1",
-			Filename: paths[0],
-			Start:    token.Position{Filename: paths[0], Offset: 0, Line: 1, Column: 1},
-			End:      token.Position{Filename: paths[0], Offset: 10, Line: 1, Column: 11},
-			Message:  "Test issue 1",
-		},
-		{
-			Rule:     "rule2",
-			Filename: paths[1],
-			Start:    token.Position{Filename: paths[1], Offset: 0, Line: 1, Column: 1},
-			End:      token.Position{Filename: paths[1], Offset: 10, Line: 1, Column: 11},
-			Message:  "Test issue 2",
-		},
-	}
-
-	mockEngine := new(mockLintEngine)
-	mockEngine.On("Run", paths[0]).Return([]types.Issue{expectedIssues[0]}, nil)
-	mockEngine.On("Run", paths[1]).Return([]types.Issue{expectedIssues[1]}, nil)
-
-	issues, err := processFiles(ctx, logger, mockEngine, paths, processFile)
-
-	assert.NoError(t, err)
-	assert.Len(t, issues, 2)
-	assert.Contains(t, issues, expectedIssues[0])
-	assert.Contains(t, issues, expectedIssues[1])
-	mockEngine.AssertExpectations(t)
-}
-
-func TestHasDesiredExtension(t *testing.T) {
-	t.Parallel()
-	assert.True(t, hasDesiredExtension("test.go"))
-	assert.True(t, hasDesiredExtension("test.gno"))
-	assert.False(t, hasDesiredExtension("test.txt"))
-	assert.False(t, hasDesiredExtension("test"))
 }
 
 func TestRunWithTimeout(t *testing.T) {
